LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/unattached-subject-inset-expected.txt


FAIL Creating a view timeline with a subject that is not attached to the document works as expected assert_equals: Null source while detached expected (object) null but got (undefined) undefined
